requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
langchain-core
langchain-text-splitters
langchain-pinecone
pinecone-client
sentence-transformers
pypdf
python-dotenv
pydantic
python-multipart